+++ opsmate/runtime/__init__.py
@@ -1,0 +1,4 @@
+from opsmate.runtime.runtime import Runtime, RuntimeError
+from opsmate.runtime.local import LocalRuntime
+
+__all__ = ["Runtime", "LocalRuntime", "RuntimeError"]
